requirements.txt
newspaper3k
lxml_html_clean
transformers
torch
pandas
pyarrow
tqdm